# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BasisGuard Evidence Log & Adaptation Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /positions:
    get:
      operationId: listPositions
      tags: [positions]
      summary: List position records
      parameters:
        - name: tier
          in: query
          schema:
            type: string
            enum: [will, should, more_likely_than_not, substantial_authority, reasonable_basis]
        - name: requires_review
          in: query
          schema:
            type: boolean
        - name: event_type
          in: query
          schema:
            type: string
        - name: classification
          in: query
          schema:
            type: string
        - name: profile_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of positions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionList"
    post:
      operationId: createPosition
      tags: [positions]
      summary: Create a new position record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionRecord"
...
  /positions/batch-signoff:
    post:
      operationId: batchSignoffPositions
      tags: [positions]
      summary: Batch sign-off on multiple positions in a single CPA attestation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchSignoffInput"
      responses:
        "200":
          description: Batch signoff result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchSignoffResult"
  /positions/tier-suggestion:
    get:
      operationId: getTierSuggestion
      tags: [positions]
      summary: Auto-suggest a confidence tier based on linked citations and event type
      parameters:
        - name: event_type
          in: query
          required: true
          schema:
            type: string
        - name: citation_ids
          in: query
          schema:
            type: array
            items:
              type: string
            default: []
      responses:
        "200":
          description: Tier suggestion with reasoning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TierSuggestion"
  /positions/{id}:
    get:
      operationId: getPosition
      tags: [positions]
      summary: Get a position record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Position record with citations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePosition
      tags: [positions]
      summary: Update a position record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionUpdate"
      responses:
        "200":
          description: Updated position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionRecord"
  /positions/{id}/signoff:
    post:
      operationId: signOffPosition
      tags: [positions]
      summary: CPA/reviewer sign-off on a position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignoffInput"
      responses:
        "200":
          description: Signed off position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionRecord"
  /positions/{id}/supersede:
    post:
      operationId: supersedePosition
      tags: [positions]
      summary: Create a new position that supersedes this one
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionInput"
      responses:
        "201":
          description: New superseding position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionRecord"
  /positions/{id}/history:
    get:
      operationId: getPositionHistory
      tags: [positions]
      summary: Get the full supersession chain for a position (oldest first)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full history chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionHistory"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Citations ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Authority Citation ─────────────────────────────────────────────────────
    AuthorityCitation:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [Notice, Rev_Proc, Rev_Rul, Treasury_Decision, Case, Statute]
        reference:
          type: string
        summary:
          type: string
        url:
          type: ["string", "null"]
        authority_strength:
          type: string
          enum: [binding_on_courts, binding_on_irs_only, non_binding_persuasive]
        created_at:
          type: string
      required:
        - id
        - type
        - reference
        - summary
        - authority_strength
        - created_at
...
    ProfileRule:
      type: object
      properties:
        event_type:
          type: string
        classification:
          type: string
        tier:
          type: string
          enum: [will, should, more_likely_than_not, substantial_authority, reasonable_basis]
        citation_ids:
          type: array
          items:
            type: string
      required:
        - event_type
        - classification
        - tier
        - citation_ids
    TreatmentProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, deprecated, opt_in_only]
        rules:
          type: array
          items:
            $ref: "#/components/schemas/ProfileRule"
        changelog:
          type: ["string", "null"]
        created_at:
          type: string
      required:
        - id
        - name
        - status
        - rules
        - created_at
...
    PositionRecord:
      type: object
      properties:
        id:
          type: string
        tx_id:
          type: ["string", "null"]
        wallet_id:
          type: ["string", "null"]
        event_type:
          type: string
        classification:
          type: string
        tier:
          type: string
          enum: [will, should, more_likely_than_not, substantial_authority, reasonable_basis]
        rationale:
          type: string
        profile_id:
          type: ["string", "null"]
        profile_version:
          type: ["string", "null"]
        requires_review:
          type: boolean
        reviewer_id:
          type: ["string", "null"]
        reviewer_name:
          type: ["string", "null"]
        reviewer_credential:
          type: ["string", "null"]
        reviewer_signoff_at:
          type: ["string", "null"]
        superseded_by:
          type: ["string", "null"]
        created_at:
          type: string
        is_stale:
          type: boolean
          description: True when a Reasonable Basis position is older than 180 days and has not been superseded
      required:
        - id
        - event_type
        - classification
        - tier
        - rationale
        - requires_review
        - created_at
        - is_stale
    PositionDetail:
      allOf:
        - $ref: "#/components/schemas/PositionRecord"
        - type: object
          properties:
            citations:
              type: array
              items:
                $ref: "#/components/schemas/AuthorityCitation"
            profile:
              $ref: "#/components/schemas/TreatmentProfile"
          required:
            - citations
    PositionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PositionRecord"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
    PositionInput:
      type: object
      properties:
        tx_id:
          type: string
        wallet_id:
          type: string
        event_type:
          type: string
        classification:
          type: string
        tier:
          type: string
          enum: [will, should, more_likely_than_not, substantial_authority, reasonable_basis]
        rationale:
          type: string
        profile_id:
          type: string
        profile_version:
          type: string
        citation_ids:
          type: array
          items:
            type: string
        requires_review:
          type: boolean
      required:
        - event_type
        - classification
        - tier
        - rationale
    PositionUpdate:
      type: object
      properties:
        rationale:
          type: string
        requires_review:
          type: boolean
        citation_ids:
          type: array
          items:
            type: string
    SignoffInput:
      type: object
      properties:
        reviewer_id:
          type: string
        reviewer_name:
          type: string
        reviewer_credential:
          type: string
      required:
        - reviewer_id
        - reviewer_name
        - reviewer_credential

    # ── Dashboard ──────────────────────────────────────────────────────────────
...
    BatchSignoffInput:
      type: object
      properties:
        position_ids:
          type: array
          items:
            type: string
          minItems: 1
        reviewer_id:
          type: string
        reviewer_name:
          type: string
        reviewer_credential:
          type: string
        note:
          type: ["string", "null"]
          description: Optional attestation note applied to all positions in the batch
      required:
        - position_ids
        - reviewer_id
        - reviewer_name
        - reviewer_credential
    BatchSignoffResult:
      type: object
      properties:
        signed_count:
          type: integer
        skipped_count:
          type: integer
          description: Positions skipped because they were already signed off
        signed_ids:
          type: array
          items:
            type: string
        skipped_ids:
          type: array
          items:
            type: string
      required:
        - signed_count
        - skipped_count
        - signed_ids
        - skipped_ids

    # ── Tier Suggestion ────────────────────────────────────────────────────────
    TierSuggestion:
      type: object
      properties:
        suggested_tier:
          type: string
          enum: [will, should, more_likely_than_not, substantial_authority, reasonable_basis]
        ceiling_tier:
          type: string
          enum: [will, should, more_likely_than_not, substantial_authority, reasonable_basis]
          description: Maximum defensible tier given the provided citations
        reasoning:
          type: string
        citation_count:
          type: integer
        strongest_authority:
          type: ["string", "null"]
        profile_rule_match:
          type: ["string", "null"]
          description: Tier from matching profile rule, if any
        flag_downgrade:
          type: boolean
          description: True when suggested tier is lower than ceiling_tier — user should be cautioned
      required:
        - suggested_tier
        - ceiling_tier
        - reasoning
        - citation_count
        - flag_downgrade

    # ── Position History ───────────────────────────────────────────────────────
    PositionHistoryEntry:
      allOf:
        - $ref: "#/components/schemas/PositionRecord"
        - type: object
          properties:
            is_current:
              type: boolean
            generation:
              type: integer
              description: 1 = original, 2 = first supersession, etc.
          required:
            - is_current
            - generation
    PositionHistory:
      type: object
      properties:
        chain_length:
          type: integer
        current_id:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/PositionHistoryEntry"
      required:
        - chain_length
        - current_id
        - entries

    # ── Export ─────────────────────────────────────────────────────────────────
...