# lib/api-zod/src/generated/types/positionHistoryEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BasisGuard Evidence Log & Adaptation Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { PositionRecord } from './positionRecord';

export type PositionHistoryEntry = PositionRecord & {
  is_current: boolean;
  /** 1 = original, 2 = first supersession, etc. */
  generation: number;
};