# lib/api-zod/src/generated/types/positionInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BasisGuard Evidence Log & Adaptation Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { PositionInputTier } from './positionInputTier';

export interface PositionInput {
  tx_id?: string;
  wallet_id?: string;
  event_type: string;
  classification: string;
  tier: PositionInputTier;
  rationale: string;
  profile_id?: string;
  profile_version?: string;
  citation_ids?: string[];
  requires_review?: boolean;
}