# lib/api-zod/src/generated/types/positionUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BasisGuard Evidence Log & Adaptation Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface PositionUpdate {
  rationale?: string;
  requires_review?: boolean;
  citation_ids?: string[];
}