# lib/api-zod/src/generated/types/signoffInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BasisGuard Evidence Log & Adaptation Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface SignoffInput {
  reviewer_id: string;
  reviewer_name: string;
  reviewer_credential: string;
}